requirements.txt
python-telegram-bot[rate-limiter,job-queue]==21.6
python-dotenv==1.0.1
httpx==0.27.2
